Moving issymbolic to smtlib
Performance optimization avoiding runtime imports
