+++ manticore/core/smtlib/__init__.py
@@ -1,4 +1,4 @@
-from .expression import Expression, Bool, BitVec, Array, BitVecConstant  # noqa
+from .expression import Expression, Bool, BitVec, Array, BitVecConstant, issymbolic  # noqa
 from .constraints import ConstraintSet  # noqa
 from .solver import *  # noqa
 from . import operators as Operators  # noqa
