+++ manticore/core/smtlib/expression.py
@@ -23,6 +23,17 @@ class Expression:
     def taint(self):
         return self._taint
 
+def issymbolic(value):
+    """
+        Helper to determine whether an object is symbolic (e.g checking
+        if data read from memory is symbolic)
+
+        :param object value: object to check
+        :return: whether `value` is symbolic
+        :rtype: bool
+        """
+    return isinstance(value, Expression)
+
 
 class Variable(Expression):
     def __init__(self, name, *args, **kwargs):
